requirements.txt
asgiref==3.2.10
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.3
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.1.1
django-widget-tweaks==1.4.8
html5lib==1.1
Pillow==2.2.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyMySQL==0.10.0
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2020.1
reportlab==3.5.49
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.4
